m365server/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via redis
azure-core==1.29.4
    # via
    #   -r requirements/requirements.in
    #   azure-storage-blob
azure-storage-blob==12.18.1
    # via -r requirements/requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via uvicorn
contourpy==1.1.1
    # via matplotlib
cryptography==41.0.3
    # via
    #   azure-storage-blob
    #   msal
    #   pyjwt
cycler==0.11.0
    # via matplotlib
debugpy==1.8.0
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.3
    # via anyio
fastapi==0.103.1
    # via -r requirements/requirements.in
fonttools==4.42.1
    # via matplotlib
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via -r requirements/requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
isodate==0.6.1
    # via azure-storage-blob
kiwisolver==1.4.5
    # via matplotlib
loguru==0.7.2
    # via -r requirements/requirements.in
matplotlib==3.8.0
    # via -r requirements/requirements.in
msal==1.24.0
    # via -r requirements/requirements.in
numpy==1.25.2
    # via
    #   -r requirements/requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   pyarrow
openpyxl==3.1.2
    # via -r requirements/requirements.in
packaging==23.1
    # via matplotlib
pandas==2.1.0
    # via -r requirements/requirements.in
pillow==10.0.1
    # via matplotlib
pyarrow==13.0.0
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pydantic==2.3.0
    # via fastapi
pydantic-core==2.6.3
    # via pydantic
pyjwt[crypto]==2.8.0
    # via msal
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via -r requirements/requirements.in
python-multipart==0.0.6
    # via -r requirements/requirements.in
pytz==2023.3.post1
    # via pandas
redis==5.0.0
    # via -r requirements/requirements.in
requests==2.31.0
    # via
    #   -r requirements/requirements.in
    #   azure-core
    #   msal
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.20
    # via -r requirements/requirements.in
starlette==0.27.0
    # via fastapi
tomlkit==0.12.1
    # via -r requirements/requirements.in
typing-extensions==4.7.1
    # via
    #   azure-core
    #   azure-storage-blob
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
tzdata==2023.3
    # via pandas
urllib3==2.0.4
    # via requests
uvicorn==0.23.2
    # via -r requirements/requirements.in
xlsxwriter==3.1.3
    # via -r requirements/requirements.in
